Added drush drupal build

+++ lib/docman/builders/git_builder.rb
@@ -2,7 +2,7 @@ require 'yaml'
 
 module Docman
   module Builders
-    class GitBuilder < Builder
+    class DrupalBuilder < Builder
 
       register_builder :git
 
@@ -14,10 +14,11 @@ module Docman
       def strip
         puts 'Do strip'
         FileUtils.rm_r(@info['full_build_path']) if File.directory? @info['full_build_path']
+        FileUtils.rm_r @info['temp_path'] if @info.need_rebuild? and File.directory? @info['temp_path']
         result = GitUtil.get(@info['repo'], @info['temp_path'], @info.version_type, @info.version)
         FileUtils.mkdir_p(@info['full_build_path'])
         FileUtils.cp_r(Dir["#{@info['temp_path']}/."], @info['full_build_path'])
-        FileUtils.rm_r(File.join(@info['full_build_path'], '.git'))
+        FileUtils.rm_r(File.join(@info['full_build_path'], '.git')) if File.directory?(File.join(@info['full_build_path'], '.git'))
         result
       end
